Task5

+++ Tyuiu.BiryukovAY.Sprint6.Task5.V9.Test/DataServiceTest.cs
@@ -8,49 +8,45 @@ namespace Tyuiu.BiryukovAY.Sprint6.Task5.V9.Test
         [TestMethod]
         public void TestMethod1()
         {
-            DataService ds = new DataService();
+            File.WriteAllText(testFilePath, "1.5\n0\n-2.3\n0.0\n4.7\n-0.000\n8.9\n");
+        }
 
-            string testFilePath = "TestInputFile.txt";
-            string[] testData = {
-                "-13.0", "-19.0", "-9.82", "13.0", "11.49",
-                "-9.71", "3.36", "0.0", "14.52", "16.0",
-                "13.66", "0.0", "4.13", "-0.11", "19.0",
-                "-17.36", "11.0", "11.28", "-12.0", "-12.35"
-            };
+        [TestCleanup]
+        public void TestCleanup()
+        {
+            if (File.Exists(testFilePath))
+                File.Delete(testFilePath);
+        }
 
-            File.WriteAllLines(testFilePath, testData);
-
+        [TestMethod]
+        public void ValidLoadFromDataFile()
+        {
             DataService ds = new DataService();
             double[] result = ds.LoadFromDataFile(testFilePath);
 
-            double[] expected = { 0.0, 0.0 };
-
+            double[] expected = { 1.5, 0, -2.3, 0.0, 4.7, -0.000, 8.9 };
             CollectionAssert.AreEqual(expected, result);
-
-            File.Delete(testFilePath);
         }
+
+        [TestMethod]
+        public void ValidGetZeroElements()
+        {
+            DataService ds = new DataService();
+            double[] data = { 1.5, 0, -2.3, 0.0, 4.7, -0.000, 8.9 };
+            double[] zeroElements = ds.GetZeroElements(data);
+
+            double[] expected = { 0, 0, 0 }; 
+            CollectionAssert.AreEqual(expected, zeroElements);
+        }
+
         [TestMethod]
         public void NoZeroElements()
         {
-            string testFilePath = "TestInputFile2.txt";
-            string[] testData = { "1.0", "2.0", "3.0", "4.0", "5.0" };
-
-            File.WriteAllLines(testFilePath, testData);
-
             DataService ds = new DataService();
-            double[] result = ds.LoadFromDataFile(testFilePath);
+            double[] data = { 1.5, 2.3, 4.7, 8.9 };
+            double[] zeroElements = ds.GetZeroElements(data);
 
-            Assert.AreEqual(0, result.Length);
-
-            File.Delete(testFilePath);
-        }
-
-        [TestMethod]
-        public void FileNotFound()
-        {
-            DataService ds = new DataService();
-
-            Assert.ThrowsException<Exception>(() => ds.LoadFromDataFile("NonexistentFile.txt"));
+            Assert.AreEqual(0, zeroElements.Length);
         }
     }
 }
